ShotGun Implemented but needs tweaking
The weapon component class is really quickly becoming a mess. Refactoring is first thing on the to-do list now. I want to clean it up before it becomes so crazy we can't understand what we are looking at.

Shotguns work and you can edit how many projectiles are shot at once but I need to give them more randomness. Some projectiles are still spawning over each other so I need to fix that.