Projectile set up is implemented
Now when the player spawns with a gun of a certain class the game will decide which projectile goes with that gun. We still only have the one projectile but since projectiles really only have  a few properties we can easily change that by creating new ones and just selecting from a dropdown box.

Burst fire weapons are still giving me trouble. All of the projectiles are spawning on top of one another. I have a possible solution I will try to implement later today or tomorrow.